requirements-dev.txt
pytest-django>=4.2.0,<5.*
pytest-cov~=3.0.0
requests-mock~=1.10.0